API Reference

Terra.AbstractBulkWeightingScheme — Type

Base type for bulk weighting/mixing schemes that calculate weighted mixture of material properties such as conductivities or densities.

Terra.AbstractClosureRelation — Type

abstract type AbstractClosureRelation

Base type for prognostic variable closure relations for differential equations of the form:

\[\frac{\partial g(u)}{\partial t} = F(u)\]

where F represents the RHS tendency as a function of the state variable u, and g(u) is a closure or constitutive relation that maps u to the physical units matching the tendency. Common examples in soil hydrothermal modeling are temperature-enthalpy and saturation-pressure relations.

Terra.AbstractEnergyBalanceModel — Type

abstract type AbstractEnergyBalanceModel{NF} <: Terra.AbstractModel{NF}

Base type for surface energy balance models.

Terra.AbstractGroundModel — Type

abstract type AbstractGroundModel{NF} <: Terra.AbstractModel{NF}

Base type for ground (e.g. soil and rock) models.

Terra.AbstractHydrologyModel — Type

abstract type AbstractHydrologyModel{NF} <: Terra.AbstractModel{NF}

Base type for surface hydrology models.

Terra.AbstractLandModel — Type

AbstractLandModel <: AbstractModel

Base type for full land models which couple together multiple component models.

Terra.AbstractLandProcess — Type

AbstractLandProcess{NF}

Base type for all land processes which define physics or parameterizations but are not standalone models.

Terra.AbstractModel — Type

abstract type AbstractModel{NF}

Base type for all models.

Terra.AbstractSnowModel — Type

abstract type AbstractSnowModel{NF} <: Terra.AbstractModel{NF}

Base type for snow models.

Terra.AbstractSoilModel — Type

abstract type AbstractSoilModel{NF} <: Terra.AbstractGroundModel{NF}

Base type for soil ground models.

Terra.AbstractTimeStepperCache — Type

Base type for time-stepper state caches.

Terra.AbstractVariable — Type

abstract type AbstractVariable

Base type for state variable specification.

Terra.AbstractVegetationModel — Type

abstract type AbstractVegetationModel{NF} <: Terra.AbstractModel{NF}

Base type for vegetation/carbon models.

Terra.AuxiliaryVariable — Type

AuxiliaryVariable{VD<:VarDims} <: AbstractVariable

Represents an auxiliary (sometimes called "diagnostic") variable with the given name and dims on the spatial grid.

Terra.ColumnGrid — Type

ColumnGrid{NF,RectGrid<:Grids.RectilinearGrid} <: AbstractLandGrid

Represents a set of laterally independent vertical columns with dimensions (x, y, z) where x is the column dimension, y=1 is constant, and z is the vertical axis.

Terra.ConstantSoilCarbonDenisty — Type

ConstantSoilCarbonDenisty{NF} <: AbstractSoilBiogeochemistry

Naive implementation of soil biogeochemistry that just assumes there to be a constant organic content in all soil layers.

Terra.FieldInitializers — Type

struct FieldInitializers{VarInits<:NamedTuple} <: Terra.AbstractInitializer

Initializer type that allows for direct specification of Field initializer functions.

Example:

initializer = FieldInitializers(:temperature => )

Terra.FieldInitializers — Method

FieldInitializers(; vars...)

Creates a FieldInitializers for the given variables. The first value of the pair is the name of the state variable while the second value should be a Function or callable struct of the form f(x,y,z)::Real where z is the vertical coordinate, decreasing with depth.

Example:

initializer = FieldInitializers(
    temperature = (x,y,z) -> 0.01*abs(z) + exp(z)*sin(2π*z)
)

Terra.ForwardEuler — Type

Simple forward Euler time stepping scheme.

Terra.GlobalRingGrid — Type

struct GlobalRingGrid{NF, RingGrid<:SpeedyWeather.RingGrids.AbstractGrid, RectGrid<:Oceananigans.Grids.AbstractGrid} <: Terra.AbstractLandGrid{NF}

Convenience wrapper around ColumnGrid that defines the columns as points on a global RingGrid.

Terra.HomogeneousSoil — Type

struct HomogeneousSoil{NF} <: Terra.AbstractStratigraphy{NF}

Represents a soil stratigraphy of well mixed material with homogeneous soil texture.

Terra.InverseQuadratic — Type

The inverse quadratic (or "quadratic parallel") bulk thermal conductivity formula (Cosenza et al. 2003):

\[k = [\sum_{i=1}^N θᵢ\sqrt{kᵢ}]^2\]

Cosenza, P., Guérin, R., and Tabbagh, A.: Relationship between thermal conductivity and water content of soils using numerical modelling, European Journal of Soil Science, 54, 581–588, https://doi.org/10.1046/j.1365-2389.2003.00539.x, 2003.

Terra.LUEPhotosynthesis — Type

struct LUEPhotosynthesis{NF} <: Terra.AbstractPhotosynthesis

Photosynthesis implementation from PALADYN (Willeit 2016) for C3 PFTs following the general light use efficiency model described in Haxeltine and Prentice 1996.

Authors: Maha Badri and Matteo Willeit

Properties:

τ25::Any: Value of τ at 25°C

Kc25::Any: Value of Kc at 25°C

Ko25::Any: Value of Ko at 25°C

q10_τ::Any: q10 for temperature-sensitive parameter τ

q10_Kc::Any: q10 for temperature-sensitive parameter Kc

q10_Ko::Any: q10 for temperature-sensitive parameter Ko

α_leaf::Any: Leaf albedo in PAR range

cq::Any: Conversion factor for solar radiation at 550 nm from J/m² to mol/m² [mol/J]

k_ext::Any: Extinction coefficient for radiation through vegetation

αa::Any: Fraction of PAR assimilated at ecosystem level, relative to leaf level

t_CO2_high::Any: Parameter, PFT specific

t_CO2_low::Any: Parameter, PFT specific

t_photos_high::Any: Parameter, PFT specific

t_photos_low::Any: Parameter, PFT specific

α_C3::Any: Intrinsic quantum efficiency of CO2 uptake in C3 plants

C_mass::Any: Atomic mass of carbon

θr::Any: Shape parameter

Terra.MedlynStomatalConductance — Type

struct MedlynStomatalConductance{NF} <: Terra.AbstractStomatalConductance

Stomatal conductance implementation from PALADYN (Willeit 2016) following the optimal stomatal conductance model (Medlyn et al. 2011).

Authors: Maha Badri and Matteo Willeit

Properties:

g1::Any: Parameter in optimal stomatal conductance formulation (Lin et al. 2015), PFT specific

Terra.Namespace — Type

Namespace <: AbstractVariable

Represents a new variable namespace, typically from a subcomponent of the model. It is (currently) assumed that tha name of the namesapce corresponds to a property defined on the model type.

Terra.PALADYNAutotrophicRespiration — Type

struct PALADYNAutotrophicRespiration{NF} <: Terra.AbstractAutotrophicRespiration

Autotrophic respiration implementation from PALADYN (Willeit 2016).

Authors: Maha Badri and Matteo Willeit

Properties:

cn_sapwood::Any: Sapwood parameter

cn_root::Any: Root parameter

aws::Any: Ratio of total to respiring stem carbon, Cox 2001, PFT specific

Terra.PALADYNCarbonDynamics — Type

struct PALADYNCarbonDynamics{NF} <: Terra.AbstractVegetationCarbonDynamics

Vegetation carbon dynamics implementation following PALADYN (Willeit 2016) but considering only the sum of the vegetation carbon pools. The subsequent splitting into Cleaf, Cstem, C_root is not implemented for now.

Authors: Maha Badri

Properties:

SLA::Any: Specific leaf area (Kattge et al. 2011) [m²/kgC], PFT specific

awl::Any: Allometric coefficient, modified from Cox 2001 to account for bwl=1 [kgC/m²], PFT specific

LAI_min::Any: Minimum Leaf Area Index modified from Clark et al. 2011, PFT specific

LAI_max::Any: Maximum Leaf Area Index modified from Clark et al. 2011, PFT specific

γL::Any: Leaf turnover rate (Kattge et al. 2011) [1/year], PFT specific

γR::Any: Root turnover rate [1/year], PFT specific

γS::Any: Stem turnover rate modified from Clark et al. 2011 [1/year], PFT specific

Terra.PALADYNPhenology — Type

struct PALADYNPhenology{NF} <: Terra.AbstractPhenology

Vegetation phenology implementation from PALADYN (Willeit 2016).

Authors: Maha Badri and Matteo Willeit

Properties:

Terra.PALADYNVegetationDynamics — Type

struct PALADYNVegetationDynamics{NF} <: Terra.AbstractVegetationDynamics

Vegetation dynamics implementation following PALADYN (Willeit 2016) for a single PFT based on the Lotka–Volterra approach.

Authors: Maha Badri

Properties:

ν_seed::Any: Vegetation seed fraction

γv_min::Any: Minimum vegetation disturbance rate [1/year]

Terra.PrescribedHydraulics — Type

PrescribedHydraulics{NF} <: AbstractSoilHydraulicProperties

Represents a simple case where soil hydraulic properties are directly prescribed to constant values. This is mostly provided just for testing, although it may be useful in certain cases where direct measurements of hydraulic properites are available.

Terra.SURFEXHydraulics — Type

struct SURFEXHydraulics{NF} <: Terra.AbstractSoilHydraulicProperties{NF}

SURFEX parameterization of mineral soil porosity (Masson et al. 2013).

Terra.SoilEnergyBalance — Type

struct SoilEnergyBalance{NF, HeatOperator<:Terra.AbstractHeatOperator, ThermalProps<:(SoilThermalProperties{NF})} <: Terra.AbstractSoilEnergyBalance{NF}

Properties:

operator::Terra.AbstractHeatOperator: Heat transport operator

thermal_properties::SoilThermalProperties: Soil thermal properties

Terra.SoilModel — Type

struct SoilModel{NF, GridType<:Terra.AbstractLandGrid{NF}, Stratigraphy<:Terra.AbstractStratigraphy, SoilEnergy<:Terra.AbstractSoilEnergyBalance, SoilHydrology<:Terra.AbstractSoilHydrology, Biogeochemistry<:Terra.AbstractSoilBiogeochemistry, Constants<:PhysicalConstants{NF}, BoundaryConditions<:Terra.AbstractBoundaryConditions, Initializer<:Terra.AbstractInitializer, TimeStepper<:Terra.AbstractTimeStepper} <: Terra.AbstractSoilModel{NF}

General implementation of a 1D column model of soil energy, water, and carbon transport.

Properties:

grid::Terra.AbstractLandGrid: Spatial grid type

strat::Terra.AbstractStratigraphy: Stratigraphy of the soil

energy::Terra.AbstractSoilEnergyBalance: Soil energy balance

hydrology::Terra.AbstractSoilHydrology: Soil hydrology/water balance

biogeochem::Terra.AbstractSoilBiogeochemistry: Soil biogeochemistry

constants::PhysicalConstants: Physical constants

boundary_conditions::Terra.AbstractBoundaryConditions: Boundary conditions

initializer::Terra.AbstractInitializer: State variable initializer

time_stepping::Terra.AbstractTimeStepper: Timestepping scheme

Terra.StateVariables — Type

struct StateVariables{prognames, tendnames, auxnames, subnames, closurenames, ProgVars, TendVars, AuxVars, SubVars, Closures, ClockType} <: Terra.AbstractStateVariables

Container type for all Fields corresponding to state variables defined by a model. StateVariables partitions the fields into three categories: prognostic, tendencies, and auxiliary. Prognostic variables are those which characterize the state of the system and are assigned tendencies to be integrated by the timestepper. Auxiliary fields are additional state variables derived from the prognostic state variables but which are conditionally independent of their values at the previous time step given the current prognostic state. It is worth noting that tendencies are also treated internally as auxiliary variables; however, they are assigned their own category here since they need to be handled separately by the timestepping scheme.

Terra.TemperatureEnergyClosure — Type

TemperatureEnergyClosure

Defines the constitutive relationship between temperature and the internal energy, U, of the system, i.e:

\[U(T) = T\times C(T) - L_f \theta_{wi} (1 - F(T))\]

where T is temperature, C(T) is the temperature-dependent heat capacity, L_f is the volumetric latent heat of fusion, and F(T) is the constitutive relation between temperature and the unfrozen fraction of pore water. Note that this formulation implies that the zero

Terra.VegetationModel — Type

struct VegetationModel{NF, Photosynthesis<:Terra.AbstractPhotosynthesis, StomatalConducatance<:Terra.AbstractStomatalConductance, AutotrophicRespiration<:Terra.AbstractAutotrophicRespiration, CarbonDynamics<:Terra.AbstractVegetationCarbonDynamics, VegetationDynamics<:Terra.AbstractVegetationDynamics, Phenology<:Terra.AbstractPhenology, GridType<:Terra.AbstractLandGrid{NF}, Constants<:PhysicalConstants{NF}, BoundaryConditions<:Terra.AbstractBoundaryConditions, Initializer<:Terra.AbstractInitializer, TimeStepper<:Terra.AbstractTimeStepper} <: Terra.AbstractVegetationModel{NF}

Model for natural (unmanaged) vegetation processes for a single plant functional type (PFT). Multiple PFTs can be later handled with a TiledVegetationModel type that composes multiple VegetationModels with different parameters for each PFT.

Properties:

grid::Terra.AbstractLandGrid: Spatial grid type

photosynthesis::Terra.AbstractPhotosynthesis: Photosynthesis scheme

stomatal_conductance::Terra.AbstractStomatalConductance: Stomatal conducantance scheme

autotrophic_respiration::Terra.AbstractAutotrophicRespiration: Autotrophic respiration scheme

phenology::Terra.AbstractPhenology: Phenology scheme

carbon_dynamics::Terra.AbstractVegetationCarbonDynamics: Vegetation carbon pool dynamics

vegetation_dynamics::Terra.AbstractVegetationDynamics: Vegetation population density or coverage fraction dynamics

constants::PhysicalConstants: Physical constants

boundary_conditions::Terra.AbstractBoundaryConditions: Boundary conditions

initializer::Terra.AbstractInitializer: State variable initializer

time_stepping::Terra.AbstractTimeStepper: Timestepping scheme

Terra.XY — Type

XY <: VarDims

Indicator type for variables that should be assigned a 2D (lateral only) field on their associated grid.

Terra.XYZ — Type

XYZ <: VarDims

Indicator type for variables that should be assigned a 3D field on their associated grid.

Terra.auxiliary — Method

auxiliary(name, dims)

Convenience constructor method for AuxiliaryVariable.

Terra.compute_auxiliary! — Function

compute_auxiliary!(state, model::AbstractModel)

Computes updates to all auxiliary variables based on the current prognostic state of the model.

Terra.compute_auxiliary! — Method

compute_auxiliary!(idx, state, model, process, args)

Terra.compute_f_deciduous — Method

compute_f_deciduous(phenol)

Computes f_deciduous, a factor for smooth transition between evergreen and deciduous.

Terra.compute_phen — Method

compute_phen(phenol)

Computes phen, the phenology factor.

Terra.compute_tendencies! — Function

compute_tendencies!(state, model::AbstractModel)

Computes tendencies for all prognostic state variables for model stored in the given state.

Terra.compute_tendencies! — Method

compute_tendencies!(idx, state, model, process, args)

Terra.compute_vpd — Method

compute_vpd(T_air::NF, q_air::NF, pres::NF) where NF

Computes the vapor pressure deficit from air temperature, specific humidity, and surface pressure.

Terra.compute_Λ_loc — Method

compute_Λ_loc(vegcarbon_dynamics, LAI_b)

Computes the local litterfall rate Λ_loc based on the balanced Leaf Area Index LAI_b (assuming evergreen PFTs).

Terra.compute_γv — Method

compute_γv(veg_dynamics)

Computes the disturbance rateγv.

Terra.compute_λ_NPP — Method

compute_λ_NPP(vegcarbon_dynamics, LAI_b)

Computes λ_NPP,a factor determining the partitioning of NPP between increase of vegetation carbon of the existing vegetated area and spreading of the given PFT based on the balanced Leaf Area Index LAI_b.

Terra.compute_ν_star — Method

compute_ν_star(veg_dynamics, ν)

Computes ν_star which is the maximum between the current vegetation fraction ν and the seed fraction ν_seed, to ensure that a PFT is always seeded.

Terra.create_field — Method

create_field(var, init, bcs, grid, args; kwargs...)

Initializes a Field on grid for the variable, var, using the given initializer, init, and boundary conditions, bcs. Additional arguments are passed direclty to the Field constructor. The location of the Field is determined by the specified VarDims on var.

Terra.fastmap — Method

fastmap(f::F, iter::NamedTuple...) where {F}

Same as map for NamedTuples but with guaranteed type stability. fastmap is a @generated function which unrolls calls to f into a loop-free tuple construction expression. All named tuples must have the same keys but in no particular order. The returned NamedTuple

Terra.fastmap — Method

fastmap(f::F, iter::NTuple{N,Any}...) where {F,N}

Same as map for NTuples but with guaranteed type stability. fastmap is a @generated function which unrolls calls to f into a loop-free tuple construction expression.

Terra.get_boundary_conditions — Method

get_boundary_conditions(model::AbstractModel)::AbstractBoundaryConditions

Returns the time stepping scheme associated with this model.

Terra.get_dt — Function

get_dt(timestepper::AbstractTimeStepper)

Get the current timestep size for the time stepper.

Terra.get_field_boundary_conditions — Method

get_field_boundary_conditions(bcs::AbstractBoundaryConditions, var::AbstractVariable)

Retrieve the Field boundary conditions for the corresponding variable. Defaults to returning nothing if no BC is defined.

Terra.get_field_grid — Method

get_field_grid(grid)

Retrieves the underlying numerical grid on which Fields are defined.

Terra.get_field_initializer — Method

get_field_initializer(init, var)

Retrieves the initializer for the given variable var or returns nothing if not defined.

Terra.get_field_initializer — Method

get_field_initializer(init::AbstractInitializer, var::AbstractVariable)

Terra.get_grid — Method

get_grid(model::AbstractModel)::AbstractGrid

Return the spatial grid associated with this model.

Terra.get_initializer — Method

get_initializer(model::AbstractModel)::AbstractInitializer

Returns the time stepping scheme associated with this model.

Terra.get_npoints — Method

get_npoints(spacing)

Return the number of vertical layers defined by this discretization.

Terra.get_spacing — Method

get_spacing(spacing)

Return a Vector of vertical layer thicknesses according to the given discretization.

Terra.get_time_stepping — Method

get_time_stepping(model::AbstractModel)::AbstractTimeStepper

Returns the time stepping scheme associated with this model.

Terra.heatcapacity — Method

heatcapacity(props, fracs)

Computes the bulk heat capacity of a finite volume from the given volumetric fractions.

Terra.initialize! — Function

initialize!(state, model::AbstractModel)
initialize!(state, model::AbstractModel, initializer::AbstractInitializer)

Calls initialize! on the model and its corresponding initializer. This method only needs to be implemented if initialization routines are necessary in addition to direct field/variable initializers.

Terra.is_adaptive — Function

is_adaptive(timestepper::AbstractTimeStepper)

Returns true if the given time stepper is adaptive, false otherwise.

Terra.merge_duplicates — Method

merge_duplicates(values::Tuple)

Filters out duplicates from the given tuple. Note that this method is not type stable or allocation-free!

Terra.namespace — Method

namespace(name)

Convenience constructor method for Namespace provided only for consistency.

Terra.organic_fraction — Method

organic_fraction(idx, state, bgc::ConstantSoilCarbonDenisty)

Calculate the soil organic carbon fraction at the given grid index. For ConstantSoilCarbonDenisty, this simply returns a constant parameter. Implementations of soil carbon dynamics would want to compute this based on the prognostic state of the carbon content/pool stored in the soil.

Terra.organic_fraction — Method

organic_fraction(bgc::ConstantSoilCarbonDenisty)

Calculate the organic solid fraction based on the prescribed SOC and natural porosity/density of the organic material.

Terra.organic_porosity — Method

organic_porosity(bgc::ConstantSoilCarbonDenisty)

Get the prescribed natural porosity of organic soil.

Terra.partial_pressure_CO2 — Method

partial_pressure_CO2(pres::NF, conc_co2::NF) where NF

Compute partial pressure of CO2 from surface pressure and CO2 concentration.

Terra.partial_pressure_O2 — Method

partial_pressure_O2(pres::NF) where NF

Compute partial pressure of oxygen from surface pressure.

Terra.prognostic — Method

prognostic(name, dims)

Convenience constructor method for PrognosticVariable.

Terra.run! — Method

run!(sim; steps, period, dt)

Run the simulation by steps or a period with dt timestep size (in seconds or Dates.Period).

Terra.thermalconductivity — Method

thermalconductivity(props, fracs)

Computes the bulk thermal conductivity of a finite volume from the given volumetric fractions.

Terra.timestep! — Function

timestep!(state, model::AbstractModel, timestepper::AbstractTimeStepper, [dt = nothing])

Advance the model state by one time step, or by dt units of time.

Terra.timestep! — Method

Convenience dispatch for timestep! that forwards to timestep!(state, model, get_time_stepping(model), dt).

Terra.timestep! — Method

timestep!(sim)

Advance the simulation forward by one timestep.

Terra.tuplejoin — Method

tuplejoin([x, y], z...)

Concatenates one or more tuples together.

Terra.vardims — Method

vardims(::AbstractVariable)

Retrieves the grid dimensions on which this variable is defined.

Terra.variables — Function

variables(model::AbstractModel)

Return a Tuple of AbstractVariables (i.e. PrognosticVariable, AuxiliaryVariable, etc.) defined by the model. For models that consist of one or more sub-models, variables may optionally be grouped into namespaces by returning a NamedTuple where the keys correspond to the group names.

Terra.varname — Method

varname(::AbstractVariable)
varname(::AbstractClosureRelation)
varname(::Pair{Symbol})

Retrieves the name of the given variable or closure. For closure relations, varname should return the name of the variable returned by the closure relation.