Soil processes

Warning

This page is a work in progress. If you have any questions or notice any errors, please raise an issue.

Subtypes of AbstractSoil represent a set of coupled processes affecting natural soils. The three main soil processes currently considered in Terrarium are:

Energy — heat conduction with phase change of water/ice in soil pores

Hydrology — hydrological properties and dynamics of the soil

Biogeochemistry — soil carbon cycling and biogeochemical processes

Terrarium.SoilEnergyWaterCarbon — Type

struct SoilEnergyWaterCarbon{NF, Stratigraphy<:Terrarium.AbstractStratigraphy{NF}, Energy<:Terrarium.AbstractSoilEnergyBalance{NF}, Hydrology<:Terrarium.AbstractSoilHydrology{NF}, Biogeochemistry<:Terrarium.AbstractSoilBiogeochemistry{NF}} <: Terrarium.AbstractSoil{NF}

Coupled process type that encapsulates the coupling of soil energy, water, and carbon dynamics. The stratigraphy parameterization determines how the vertical layering of the soil is parameterized.

variables(SoilEnergyWaterCarbon(Float32))

Variables
├─ Prognostic: 
├── internal_energy [J m^-3] on XYZ{Center, Center, Center}
├─ Auxiliary: 
├── temperature [°C] on XYZ{Center, Center, Center}
├── liquid_water_fraction [-] on XYZ{Center, Center, Center}
├── ground_temperature [°C] on XY{Center, Center}
├── saturation_water_ice [-] on XYZ{Center, Center, Center}
├── water_table [m] on XY{Center, Center}
├── hydraulic_conductivity [m s^-1] on XYZ{Center, Center, Face}
├─ Inputs: 
├─ Namespaces:

Process Interface

Terrarium.initialize! — Method

Initialize the soil energy, water, and carbon state variables on grid given the parameter values in constants.

Terrarium.compute_auxiliary! — Method

Compute auxiliary variables for soil energy, water, and carbon state variables on grid based on the given values in constants.

Terrarium.compute_tendencies! — Method

Compute tendencies for soil energy, water, and carbon state variables on grid based on the given values in constants.

Methods

Terrarium.get_stratigraphy — Method

get_stratigraphy(soil::AbstractSoil)

Return the stratigraphy parameterization associated with soil.

Terrarium.get_energy_balance — Method

get_energy_balance(soil::AbstractSoil)

Return the energy balance scheme associated with soil.

Terrarium.get_hydrology — Method

get_hydrology(soil::AbstractSoil)

Return the hydrology scheme associated with soil.

Terrarium.get_biogeochemistry — Method

get_biogeochemistry(soil::AbstractSoil)

Return the biogeochemistry scheme associated with soil.

Component processes

See the following pages for more details on the various implementations and relevant parameterizations for each soil process:

Soil energy balance — heat conduction and freeze-thaw

Soil hydrology — water transport and saturation

Soil stratigraphy — vertical distribution of material properties