Plant available water

Warning

This page is a work in progress. If you have any questions or notice any errors, please raise an issue.

Overview

The plant available water (PAW) represents the fraction of soil water that is available for uptake by plant roots. This quantity is critical for constraining photosynthesis and transpiration fluxes.

Terrarium.AbstractPlantAvailableWater — Type

abstract type AbstractPlantAvailableWater{NF} <: Terrarium.AbstractProcess{NF}

Base type for processes that comptue the plant available water fraction in each soil layer.

subtypes(Terrarium.AbstractPlantAvailableWater)

1-element Vector{Any}:
 FieldCapacityLimitedPAW

Soil moisture limiting factor

The impact of soil water availability on photosynthesis is aggregated across the rooting depth using the root distribution:

\[\begin{equation} \beta_s = \int_0^{z_{\text{max}}} W(z) \cdot r(z) \, dz \end{equation}\]

where $r(z)$ is the normalized root fraction at depth $z$, and $z_{\text{max}}$ is the rooting depth. This soil moisture limiting factor $\beta_s \in \[0, 1\]$ directly modulates the rate of autotrophic respiration and thus also the net primary plant productivity of vegetation.

Implementations

Field capacity limited PAW

A common assumption is to consider PAW to be the fraction of water between the wilting point (where plants can no longer extract water) and field capacity (approximately optimal water availability). The fraction of available water for a given soil layer is then defined as,

\[\begin{equation} W = \min\left(\frac{\theta_w - \theta_{wp}}{\theta_{fc} - \theta_{wp}}, 1\right) \end{equation}\]

where $\theta_w$ is the current volumetric water content, $\theta_{wp}$ is the wilting point, and $\theta_{fc}$ is the field capacity matric potential. Both wilting point and field capacity are derived from soil texture properties and the water retention characteristics of the soil.

Terrarium.FieldCapacityLimitedPAW — Type

struct FieldCapacityLimitedPAW{NF} <: Terrarium.AbstractPlantAvailableWater{NF}

Implementation of vegetation water availability (a.k.a "plant available water") that computes the wilting fraction

\[W_i = \min\left(\frac{\theta_{\text{w},i} - \theta_{\text{wp},i}}{\theta_{\text{fc},i} - \theta_{\text{wp},i}} , 1\right)\]

where $\theta_{\text{w},i}$ is the volumetric water content of the $i$'th soil layer, $\theta_{\text{fc},i}$ is the "field capacity", and $\theta_{\text{wp},i}$ is the "wilting point". The water availability

Properties:

variables(FieldCapacityLimitedPAW(Float32))

Variables
├─ Prognostic: 
├─ Auxiliary: 
├── plant_available_water [-] on XYZ{Center, Center, Center}
├── soil_moisture_limiting_factor [-] on XY{Center, Center}
├─ Inputs: 
├── root_fraction [-] on XYZ{Center, Center, Center}
├─ Namespaces:

Process interface

Terrarium.compute_auxiliary! — Method

compute_auxiliary!(
    state,
    grid,
    paw::FieldCapacityLimitedPAW,
    soil::Terrarium.AbstractSoil,
    args...
) -> Any

Methods

Terrarium.soil_moisture_limiting_factor — Function

soil_moisture_limiting_factor(
    _::FieldCapacityLimitedPAW,
    grid,
    clock,
    fields
) -> Field{LX, LY, LZ, O, G, I, D, T, B, Oceananigans.Fields.FieldStatus{Float64}} where {LX, LY, LZ, O, G, I, D, T, B}

Field constructor for the soil moisture limiting factor. Returns a derived Field that calculates the integral of W(z) * r(z) where W is the water availability coefficient and r is the root fraction.

Kernel functions

Terrarium.compute_plant_available_water — Function

compute_plant_available_water(
    i, j, k, grid, fields,
    paw::AbstractPlantAvailableWater,
    soil::AbstractSoil
)

Comptue the plant-available water fraction for grid cell i, j and soil layer k.

Terrarium.compute_plant_available_water! — Function

compute_plant_available_water!(
    out,
    i,
    j,
    k,
    grid,
    fields,
    paw::FieldCapacityLimitedPAW{NF},
    strat::Terrarium.AbstractStratigraphy,
    hydrology::Terrarium.AbstractSoilHydrology,
    bgc::Terrarium.AbstractSoilBiogeochemistry,
    args...
) -> Any

Mutating wrapper for compute_plant_available_water that stores the result in out.