API Reference

Core.Type — Method

Convenience constructor for all AbstractLandModel types that allows the grid to be passed as the first positional argument.

Oceananigans.Fields.Field — Type

Field(
    grid::AbstractLandGrid,
    dims::VarDims,
    boundary_conditions=nothing,
    args...;
    kwargs...
)

Auxiliary constructor for an Oceananigans Field on grid with the given Terrarium variable dims and boundary conditions. Additional arguments are passed direclty to the Field constructor. The location of the Field is determined by VarDims defined on var.

RingGrids.Field — Method

Field(field, grid; fill_value)

Converts the given Oceananigans Field to a RingGrids.Field with a ring grid matching that of

Terrarium.AbstractBulkWeightingScheme — Type

Base type for bulk weighting/mixing schemes that calculate weighted mixture of material properties such as conductivities or densities.

Terrarium.AbstractClosureRelation — Type

abstract type AbstractClosureRelation

Base type for prognostic variable closure relations for differential equations of the form:

\[\frac{\partial g(u)}{\partial t} = F(u)\]

where F represents the RHS tendency as a function of the state variable u, and g(u) is a closure or constitutive relation that maps u to the physical units matching the tendency. Common examples in soil hydrothermal modeling are temperature-enthalpy and saturation-pressure relations.

Terrarium.AbstractEnergyBalanceModel — Type

abstract type AbstractEnergyBalanceModel{NF, GR, TS} <: Terrarium.AbstractModel{NF, GR, TS}

Base type for surface energy balance models.

Terrarium.AbstractGroundModel — Type

abstract type AbstractGroundModel{NF, GR, TS} <: Terrarium.AbstractModel{NF, GR, TS}

Base type for ground (e.g. soil and rock) models.

Terrarium.AbstractHydrologyModel — Type

abstract type AbstractHydrologyModel{NF, GR, TS} <: Terrarium.AbstractModel{NF, GR, TS}

Base type for surface hydrology models.

Terrarium.AbstractInitializer — Type

Base type for model initializers. Implementations should provide a dispatch of the initialize!(state, model::M, init::I) method where M corresponds to the model type and I to the initializer. An implementation of get_field_initializers can also be provided which returns a NamedTuple of initializer functions for individual state variable fields.

Terrarium.AbstractInteraction — Type

AbstractInteraction

Base type for interactions between models or processes.

Terrarium.AbstractLandModel — Type

AbstractLandModel <: AbstractModel

Base type for full land models which couple together multiple component models.

Terrarium.AbstractModel — Type

abstract type AbstractModel{NF, Grid<:(Terrarium.AbstractLandGrid{NF}), TS<:Terrarium.AbstractTimeStepper}

Base type for all models.

Terrarium.AbstractProcess — Type

AbstractProcess{NF}

Base type for all processes which define physics or parameterizations but are not standalone models.

Terrarium.AbstractSnowModel — Type

abstract type AbstractSnowModel{NF, GR, TS} <: Terrarium.AbstractModel{NF, GR, TS}

Base type for snow models.

Terrarium.AbstractSoilModel — Type

abstract type AbstractSoilModel{NF, GR, TS} <: Terrarium.AbstractGroundModel{NF, GR, TS}

Base type for soil ground models.

Terrarium.AbstractTimeStepperCache — Type

Base type for time-stepper state caches.

Terrarium.AbstractVariable — Type

abstract type AbstractVariable{VD<:Terrarium.VarDims}

Base type for state variable specification.

Terrarium.AbstractVegetationModel — Type

abstract type AbstractVegetationModel{NF, GR, TS} <: Terrarium.AbstractModel{NF, GR, TS}

Base type for vegetation/carbon models.

Terrarium.AnyField — Type

Type alias for an AbstractField with any X, Y, Z location or grid.

Terrarium.AnyFieldTimeSeries — Type

Type alias for a FieldTimeSeries with any X, Y, Z location or grid.

Terrarium.AuxiliaryVariable — Type

struct AuxiliaryVariable{VD<:Terrarium.VarDims, UT<:Unitful.Units} <: Terrarium.AbstractVariable{VD<:Terrarium.VarDims}

Represents an auxiliary (sometimes called "diagnostic") variable with the given name and dims on the spatial grid.

Terrarium.ColumnBCs — Type

Alias for ColumnBoundaryConditions

Terrarium.ColumnBoundaryConditions — Type

struct ColumnBoundaryConditions{TopBC, BottomBC} <: Terrarium.AbstractBoundaryConditions

Represents the boundary conditions applied at the top and bottom of a 1D column model discrietized along the vertical (depthwise) axis.

Properties:

top::Any: Boundary condition(s) applied at the top of the vertical column.

bottom::Any: Boundary condition(s) applied at the botom of the vertical column.

Terrarium.ColumnGrid — Type

ColumnGrid{NF, Arch<:AbstractArchitecture, RectGrid<:OceananigansGrids.RectilinearGrid} <: AbstractLandGrid

Represents a set of laterally independent vertical columns with dimensions (x, y, z) where x is the column dimension, y=1 is constant, and z is the vertical axis.

Terrarium.ColumnRingGrid — Type

struct ColumnRingGrid{NF, Arch<:Oceananigans.Architectures.AbstractArchitecture, RingGrid<:RingGrids.AbstractGrid, MaskField<:(RingGrids.AbstractField{Bool}), RectGrid<:Oceananigans.Grids.RectilinearGrid} <: Terrarium.AbstractLandGrid{NF, Arch<:Oceananigans.Architectures.AbstractArchitecture}

Represents a global (spherical) grid of independent, vertical columns where the spatial discretization in the horizontal direction is defined by a RingGrids.AbstractGrid. The resulting

Terrarium.ConstantSoilCarbonDenisty — Type

struct ConstantSoilCarbonDenisty{NF} <: Terrarium.AbstractSoilBiogeochemistry{NF}

Naive implementation of soil biogeochemistry that just assumes there to be a constant organic content in all soil layers.

Properties:

ρ_soc::Any: Soil organic carbon density [kg/m^3]

ρ_org::Any: Pure organic matter density [kg/m^3]

por_org::Any: Natural porosity of organic material

Terrarium.DefaultBoundaryConditions — Type

Represents default boundary conditions for all state variables. This typically implies zero-flux boundary conditions on bounded domains and periodic boundary conditions on periodic domains.

Terrarium.DefaultInitializer — Type

Marker type for a no-op initializer that leaves all Fields set to their default values.

Terrarium.FieldBCs — Type

Alias for a NamedTuple of BoundaryCondition types.

Terrarium.FieldInitializers — Type

Creates a new FieldInitializers from the given keyword arguments where each argument corresponds to an initializer function or value for specific state variable defined in the model. The initializers can be any function, value, array, or Field that would be a valid input x to Oceananigans.set!(field, x). Optionally, FieldInitializers can wrap another

FieldInitializers(temperature=(x,z) -> sin(2π*z), pore_water_ice_saturation=0)

Terrarium.FieldInitializers — Type

struct FieldInitializers{names, Init<:Terrarium.AbstractInitializer, FieldInits} <: Terrarium.AbstractInitializer

Container type that bundles one or more initializer functions for individual state variable fields into a single AbstractInitializer. FieldInitializers can also optionally wrap another AbstractInitializer type whose field initializers will be merged with those in vars.

Terrarium.FieldInputSource — Type

struct FieldInputSource{NF, VD<:Terrarium.VarDims, names, Fields<:Tuple{Vararg{Oceananigans.Fields.AbstractField{LX, LY, LZ, G, NF} where {LX, LY, LZ, G}}}} <: InputSource{NF}

Input source that reads directly from pre-specified Oceananigans Fields.

Terrarium.FieldTimeSeriesInputSource — Type

struct FieldTimeSeriesInputSource{NF, VD<:Terrarium.VarDims, names, FTS<:Tuple{Vararg{FieldTimeSeries{LX, LY, LZ, TI, K, I, D, G, NF} where {LX, LY, LZ, TI, K, I, D, G}}}} <: InputSource{NF}

Input source that reads input fields from pre-specified Oceananigans FieldTimeSeries.

Terrarium.ForwardEuler — Type

Simple forward Euler time stepping scheme.

Terrarium.HomogeneousSoil — Type

struct HomogeneousSoil{NF} <: Terrarium.AbstractStratigraphy{NF}

Represents a soil stratigraphy of well mixed material with homogeneous soil texture.

Properties:

texture::SoilTexture: Material composition of mineral soil componnet

Terrarium.Input — Type

Type-stable and GPU-friendly placeholder for input variables in model/process structs, allowing parameters/constants to be easily replaced by input Fields.

Terrarium.InputFields — Type

struct InputFields{NF, Grid<:(Terrarium.AbstractLandGrid{NF}), Input2D<:Oceananigans.Fields.AbstractField, Input3D<:Oceananigans.Fields.AbstractField}

Container type for holding data Fields for input variables. InputFields stores the Fields for each variable in separate dictionaries for 2D and 3D variables defined on the given grid. This allows for input variable Fields to be lazily allocated as-needed.

Terrarium.InputProvider — Type

struct InputProvider{NF, Grid<:(Terrarium.AbstractLandGrid{NF}), Inputs<:(InputFields{NF, Grid<:(Terrarium.AbstractLandGrid{NF})}), SourceType<:InputSource}

Acts as an interface between a given InputFields container and one or more InputSources that are responsible for initializing and updating the input Fields at each time step.

Terrarium.InputSource — Type

abstract type InputSource{NF}

Base type for input data sources. Implementations of InputSource are free to load data from any arbitrary backend but are required to implement the update_inputs!(::InputFields, ::InputSource, ::Clock) method. Implementations should additionally provide a constructor as a dispatch of InputSource.

The type argument NF corresponds to the numeric type of the input data.

Terrarium.InputVariable — Type

struct InputVariable{VD<:Terrarium.VarDims, UT<:Unitful.Units} <: Terrarium.AbstractVariable{VD<:Terrarium.VarDims}

Represents an input (e.g. forcing) variable with the given name and dims on the spatial grid.

Terrarium.InverseQuadratic — Type

The inverse quadratic (or "quadratic parallel") bulk thermal conductivity formula (Cosenza et al. 2003):

\[k = [\sum_{i=1}^N θᵢ\sqrt{kᵢ}]^2\]

Cosenza, P., Guérin, R., and Tabbagh, A.: Relationship between thermal conductivity and water content of soils using numerical modelling, European Journal of Soil Science, 54, 581–588, https://doi.org/10.1046/j.1365-2389.2003.00539.x, 2003.

Terrarium.LUEPhotosynthesis — Type

struct LUEPhotosynthesis{NF} <: Terrarium.AbstractPhotosynthesis

Photosynthesis implementation from PALADYN (Willeit 2016) for C3 PFTs following the general light use efficiency model described in Haxeltine and Prentice 1996.

Authors: Maha Badri and Matteo Willeit

Properties:

τ25::Any: Value of τ at 25°C

Kc25::Any: Value of Kc at 25°C

Ko25::Any: Value of Ko at 25°C

q10_τ::Any: q10 for temperature-sensitive parameter τ

q10_Kc::Any: q10 for temperature-sensitive parameter Kc

q10_Ko::Any: q10 for temperature-sensitive parameter Ko

α_leaf::Any: Leaf albedo in PAR range [-]

cq::Any: Conversion factor for solar radiation at 550 nm from J/m² to mol/m² [mol/J]

k_ext::Any: Extinction coefficient for radiation through vegetation [-]

α_a::Any: Fraction of PAR assimilated at ecosystem level, relative to leaf level [-]

t_CO2_high::Any: Parameter, PFT specific [°C]

t_CO2_low::Any: Parameter, PFT specific [°C]

t_photos_high::Any: Parameter, PFT specific [°C]

t_photos_low::Any: Parameter, PFT specific [°C]

α_C3::Any: Intrinsic quantum efficiency of CO2 uptake in C3 plants [mol/mol]

C_mass::Any: Atomic mass of carbon [gC/mol]

θ_r::Any: Shape parameter [-]

day_length::Any: Day length [h/day]

sec_day::Any: Seconds per day [s/day]

Terrarium.MedlynStomatalConductance — Type

struct MedlynStomatalConductance{NF} <: Terrarium.AbstractStomatalConductance

Stomatal conductance implementation from PALADYN (Willeit 2016) following the optimal stomatal conductance model (Medlyn et al. 2011).

Authors: Maha Badri and Matteo Willeit

Properties:

g1::Any: Parameter in optimal stomatal conductance formulation, Lin et al. 2015 [-], PFT specific

Terrarium.ModelState — Type

struct ModelState{NF, Arch<:Oceananigans.Architectures.AbstractArchitecture, Grid<:Terrarium.AbstractLandGrid{NF, Arch<:Oceananigans.Architectures.AbstractArchitecture}, TimeStepper<:Terrarium.AbstractTimeStepper, Model<:Terrarium.AbstractModel{NF, Grid<:Terrarium.AbstractLandGrid{NF, Arch<:Oceananigans.Architectures.AbstractArchitecture}, TimeStepper<:Terrarium.AbstractTimeStepper}, TimeStepperCache<:Terrarium.AbstractTimeStepperCache, StateVars<:Terrarium.AbstractStateVariables, Inputs<:InputProvider} <: Oceananigans.AbstractModel{TimeStepper<:Terrarium.AbstractTimeStepper, Arch<:Oceananigans.Architectures.AbstractArchitecture}

Represents the state of a "simulation" for a given model. ModelState consists of a clock, a model, and an initialized StateVariables data structure, as well as a cache for the timestepper and any relevant inputs provided by a corresponding InputProvider. The ModelState implements the Oceananigans.AbstractModel interface and can thus be provided as a "model" to all Oceananigans types.

Terrarium.Namespace — Type

struct Namespace{Vars}

Represents a new variable namespace, typically from a subcomponent of the model. It is (currently) assumed that tha name of the namespace corresponds to a property defined on the model type.

Terrarium.PALADYNAutotrophicRespiration — Type

struct PALADYNAutotrophicRespiration{NF} <: Terrarium.AbstractAutotrophicRespiration

Autotrophic respiration implementation from PALADYN (Willeit 2016).

Authors: Maha Badri and Matteo Willeit

Properties:

cn_sapwood::Any: Sapwood parameter

cn_root::Any: Root parameter

aws::Any: Ratio of total to respiring stem carbon, Cox 2001, PFT specific [-]

Terrarium.PALADYNCarbonDynamics — Type

struct PALADYNCarbonDynamics{NF} <: Terrarium.AbstractVegetationCarbonDynamics

Vegetation carbon dynamics implementation following PALADYN (Willeit 2016) but considering only the sum of the vegetation carbon pools. The subsequent splitting into Cleaf, Cstem, C_root is not implemented for now.

Authors: Maha Badri

Properties:

SLA::Any: Specific leaf area (Kattge et al. 2011) [m²/kgC], PFT specific

awl::Any: Allometric coefficient, modified from Cox 2001 to account for bwl=1 [kgC/m²], PFT specific

LAI_min::Any: Minimum Leaf Area Index modified from Clark et al. 2011 [m²/m²], PFT specific

LAI_max::Any: Maximum Leaf Area Index modified from Clark et al. 2011 [m²/m²], PFT specific

γL::Any: Leaf turnover rate (Kattge et al. 2011) [1/year], PFT specific

γR::Any: Root turnover rate [1/year], PFT specific

γS::Any: Stem turnover rate modified from Clark et al. 2011 [1/year], PFT specific

Terrarium.PALADYNPhenology — Type

struct PALADYNPhenology{NF} <: Terrarium.AbstractPhenology

Vegetation phenology implementation from PALADYN (Willeit 2016).

Authors: Maha Badri and Matteo Willeit

Properties:

Terrarium.PALADYNVegetationDynamics — Type

struct PALADYNVegetationDynamics{NF} <: Terrarium.AbstractVegetationDynamics

Vegetation dynamics implementation following PALADYN (Willeit 2016) for a single PFT based on the Lotka–Volterra approach.

Authors: Maha Badri

Properties:

ν_seed::Any: Vegetation seed fraction [-]

γv_min::Any: Minimum vegetation disturbance rate [1/year]

Terrarium.PhysicalConstants — Type

struct PhysicalConstants{NF}

A collection of general physical constants that do not (usually) need to be varied in parameter calibration.

Terrarium.PrescribedAtmosphere — Type

Terrarium.PrescribedBC — Type

struct PrescribedBC{progvar, BC, FT} <: Terrarium.AbstractBoundaryConditions

Container type for an Oceananigans BoundaryCondition applied to prognostic (or closure) variable progvar.

Properties:

value::Any: Boundary condition value

Terrarium.PrescribedHydraulics — Type

PrescribedHydraulics{NF} <: AbstractSoilHydraulicProperties

Represents a simple case where soil hydraulic properties are directly prescribed to constant values. This is mostly provided just for testing, although it may be useful in certain cases where direct measurements of hydraulic properites are available.

Properties:

cond_sat::Any: Hydraulic conductivity at saturation [m/s]

porosity::Any: Prescribed soil porosity [-]

field_capacity::Any: Prescribed field capacity [-]

wilting_point::Any: Prescribed wilting point [-]

Terrarium.PrognosticVariable — Type

struct PrognosticVariable{VD<:Terrarium.VarDims, UT<:Unitful.Units, TV<:Union{Nothing, AuxiliaryVariable}, CL<:Union{Nothing, Terrarium.AbstractClosureRelation}} <: Terrarium.AbstractVariable{VD<:Terrarium.VarDims}

Represents a prognostic variable with the given name and dims on the spatial grid.

Terrarium.SURFEXHydraulics — Type

struct SURFEXHydraulics{NF} <: Terrarium.AbstractSoilHydraulicProperties{NF}

SURFEX parameterization of mineral soil porosity (Masson et al. 2013).

Terrarium.SoilBC — Type

struct SoilBC{EnergyBC, WaterBC} <: Terrarium.AbstractBoundaryConditions

Boundary condition type for soil models that provides boundary conditions for each of the relevant internal processes, i.e. energy, hydrology, ...

Terrarium.SoilEnergyBalance — Type

struct SoilEnergyBalance{NF, HeatOperator<:Terrarium.AbstractHeatOperator, ThermalProps<:(SoilThermalProperties{NF})} <: Terrarium.AbstractSoilEnergyBalance{NF}

Standard implementation of the soil energy balance accounting for freezing and thawing of pore water/ice.

Properties:

operator::Terrarium.AbstractHeatOperator: Heat transport operator

thermal_properties::SoilThermalProperties: Soil thermal properties

Terrarium.SoilHeatCapacities — Type

struct SoilHeatCapacities{NF}

Properties:

water::Any

ice::Any

air::Any

mineral::Any

organic::Any

Terrarium.SoilHydrology — Type

struct SoilHydrology{NF, SoilWaterFluxes<:Terrarium.AbstractSoilWaterFluxes, SoilHydraulicProperties<:Terrarium.AbstractSoilHydraulicProperties{NF}, FC<:FreezeCurves.FreezeCurve} <: Terrarium.AbstractSoilHydrology{NF}

Properties:

fluxes::Terrarium.AbstractSoilWaterFluxes: Soil water flux scheme

hydraulic_properties::Terrarium.AbstractSoilHydraulicProperties: Soil hydraulic properties parameterization

freezecurve::FreezeCurves.FreezeCurve: Soil freezing and water retention curve(s) from FreezeCurves.jl

Terrarium.SoilModel — Type

struct SoilModel{NF, GridType<:(Terrarium.AbstractLandGrid{NF}), Stratigraphy<:Terrarium.AbstractStratigraphy, SoilEnergy<:Terrarium.AbstractSoilEnergyBalance, SoilHydrology<:Terrarium.AbstractSoilHydrology, Biogeochemistry<:Terrarium.AbstractSoilBiogeochemistry, Constants<:PhysicalConstants{NF}, BoundaryConditions<:Terrarium.AbstractBoundaryConditions, Initializer<:Terrarium.AbstractInitializer, TimeStepper<:Terrarium.AbstractTimeStepper} <: Terrarium.AbstractSoilModel{NF, GridType<:(Terrarium.AbstractLandGrid{NF}), TimeStepper<:Terrarium.AbstractTimeStepper}

General implementation of a 1D column model of soil energy, water, and carbon transport.

Properties:

grid::Terrarium.AbstractLandGrid: Spatial grid type

strat::Terrarium.AbstractStratigraphy: Stratigraphy of the soil

energy::Terrarium.AbstractSoilEnergyBalance: Soil energy balance

hydrology::Terrarium.AbstractSoilHydrology: Soil hydrology/water balance

biogeochem::Terrarium.AbstractSoilBiogeochemistry: Soil biogeochemistry

constants::PhysicalConstants: Physical constants

boundary_conditions::Terrarium.AbstractBoundaryConditions: Boundary conditions

initializer::Terrarium.AbstractInitializer: State variable initializer

time_stepping::Terrarium.AbstractTimeStepper: Timestepping scheme

Terrarium.SoilTexture — Type

SoilTexture{NF}

Represents soil texture as a fractional mixture of sand, silt, and clay.

Terrarium.SoilThermalConductivities — Type

struct SoilThermalConductivities{NF}

Properties:

water::Any

ice::Any

air::Any

mineral::Any

organic::Any

Terrarium.SoilThermalProperties — Type

struct SoilThermalProperties{NF, CondWeighting}

Properties:

cond::SoilThermalConductivities: Thermal conductivities for all constituents

cond_bulk::Any: Thermal conductivity mixing scheme

heatcap::SoilHeatCapacities: Thermal conductivities for all constituents

Terrarium.StateVariables — Type

struct StateVariables{prognames, tendnames, auxnames, inputnames, nsnames, closurenames, ProgFields, TendFields, AuxFields, InputFields, Namespaces, Closures, ClockType} <: Terrarium.AbstractStateVariables

Container type for all Fields corresponding to state variables defined by a model. StateVariables partitions the fields into three categories: prognostic, tendencies, and auxiliary. Prognostic variables are those which characterize the state of the system and are assigned tendencies to be integrated by the timestepper. Auxiliary fields are additional state variables derived from the prognostic state variables but which are conditionally independent of their values at the previous time step given the current prognostic state. It is worth noting that tendencies are also treated internally as auxiliary variables; however, they are assigned their own category here since they need to be handled separately by the timestepping scheme.

Terrarium.TemperatureEnergyClosure — Type

TemperatureEnergyClosure

Defines the constitutive relationship between temperature and the internal energy, U, of the system, i.e:

\[U(T) = T\times C(T) - L_f \theta_{wi} (1 - F(T))\]

where T is temperature, C(T) is the temperature-dependent heat capacity, L_f is the volumetric latent heat of fusion, and F(T) is the constitutive relation between temperature and the unfrozen fraction of pore water. Note that this formulation implies that the zero

Terrarium.TracerGas — Type

Generic type representing the concentration of a particular tracer gas in the atmosphere.

Terrarium.VegetationModel — Type

struct VegetationModel{NF, Photosynthesis<:Terrarium.AbstractPhotosynthesis, StomatalConducatance<:Terrarium.AbstractStomatalConductance, AutotrophicRespiration<:Terrarium.AbstractAutotrophicRespiration, CarbonDynamics<:Terrarium.AbstractVegetationCarbonDynamics, VegetationDynamics<:Terrarium.AbstractVegetationDynamics, Phenology<:Terrarium.AbstractPhenology, GridType<:(Terrarium.AbstractLandGrid{NF}), Constants<:PhysicalConstants{NF}, BoundaryConditions<:Terrarium.AbstractBoundaryConditions, Initializer<:Terrarium.AbstractInitializer, TimeStepper<:Terrarium.AbstractTimeStepper} <: Terrarium.AbstractVegetationModel{NF, GridType<:(Terrarium.AbstractLandGrid{NF}), TimeStepper<:Terrarium.AbstractTimeStepper}

Model for natural (unmanaged) vegetation processes for a single plant functional type (PFT). Multiple PFTs can be later handled with a TiledVegetationModel type that composes multiple VegetationModels with different parameters for each PFT.

Properties:

grid::Terrarium.AbstractLandGrid: Spatial grid type

photosynthesis::Terrarium.AbstractPhotosynthesis: Photosynthesis scheme

stomatal_conductance::Terrarium.AbstractStomatalConductance: Stomatal conducantance scheme

autotrophic_respiration::Terrarium.AbstractAutotrophicRespiration: Autotrophic respiration scheme

phenology::Terrarium.AbstractPhenology: Phenology scheme

carbon_dynamics::Terrarium.AbstractVegetationCarbonDynamics: Vegetation carbon pool dynamics

vegetation_dynamics::Terrarium.AbstractVegetationDynamics: Vegetation population density or coverage fraction dynamics

constants::PhysicalConstants: Physical constants

boundary_conditions::Terrarium.AbstractBoundaryConditions: Boundary conditions

initializer::Terrarium.AbstractInitializer: State variable initializer

time_stepping::Terrarium.AbstractTimeStepper: Timestepping scheme

Terrarium.XY — Type

XY <: VarDims

Indicator type for variables that should be assigned a 2D (lateral only) field on their associated grid.

Terrarium.XYZ — Type

XYZ <: VarDims

Indicator type for variables that should be assigned a 3D field on their associated grid.

Oceananigans.Simulations.run! — Method

run!(state; steps, period, dt)

Run the simulation by steps or a period with dt timestep size (in seconds or Dates.Period).

Terrarium.AmbientCO2 — Method

Creates a TracerGas for ambient CO2 with a prescribed concentration conc.

Terrarium.ConstantInitialSoilTemperature — Method

Creates a constant soil temperature initializer.

Terrarium.FreeDrainage — Method

Alias for PrescribedGradient representing a Neumann-type zero pressure gradient at the bottom of the soil column, thereby allowing free drainage of water.

Terrarium.GeothermalHeatFlux — Method

Alias for PrescribedFlux with name Q_geo representing the geothermal heat flux at the bottom boundary of the soil column.

Terrarium.GroundHeatFlux — Method

Alias for PrescribedFlux with name Q_g representing the net ground heat flux at the soil surface.

Terrarium.ImpermeableBoundary — Method

Alias for NoFlux representing a zero-flux boundary condition for water flow (prognostic variable pore_water_ice_saturation).

Terrarium.InfiltrationFlux — Method

Alias for PrescribedFlux with name Q_inf representing liquid water infiltration at the soil surface.

Terrarium.PiecewiseLinearInitialSoilTemperature — Method

Creates a piecwise linear temperature initializer from the given knots.

initializer = PiecewiseLinearInitialSoilTemperature(
    0.0u"m" => 5.0, # always in °C!
    0.5u"m" => 2.0,
    1.0u"m" => 1.0,
    10.0u"m" => 1.5,
    ...
)

Terrarium.QuasiThermalSteadyState — Method

Computes a linear temperature profile in quasi-steady state based on the given surface temperature, geothermal heat flux, and bulk thermal conductivity.

Terrarium.SoilBoundaryConditions — Method

Alias for ColumnBoundaryConditions with defaults suitable for SoilModels.

Terrarium.TracerGases — Method

Creates a NamedTuple from the given tracer gas types.

Terrarium.auxiliary — Method

auxiliary(name, dims; units, desc)

Convenience constructor method for AuxiliaryVariable.

Terrarium.compute_APAR — Method

compute_APAR(photo, swdown, LAI)

Computes absorbed PAR limited by the fraction of PAR assimilated at ecosystem level APAR [mol/m²/day], Eq. 62, PALADYN (Willeit 2016).

Terrarium.compute_Ag — Method

compute_Ag(photo, c_1, c_2, APAR, Vc_max, β)

Computes the daily gross photosynthesis Ag [gC/m²/day], Eqn 2, Haxeltine & Prentice 1996

Terrarium.compute_And — Method

compute_And(photo, c_1, c_2, APAR, Vc_max, β, Rd)

Computes the total daytime net photosynthesis And [gC/m²/day], Eqn 19, Haxeltine & Prentice 1996 + Eq. 65, PALADYN (Willeit 2016).

Terrarium.compute_C_veg_tend — Method

compute_C_veg_tend(vegcarbon_dynamics, LAI_b, NPP)

Computes the C_veg tendency based on NPP and the balanced Leaf Area Index LAI_b, Eq. 72, PALADYN (Willeit 2016)

Terrarium.compute_JE_JC — Method

compute_JE_JC(photo, c_1, c_2, APAR, Vc_max)

Computes the PAR-limited and the rubisco-activity-limited photosynthesis rates JE and JC [gC/m²/day], Eqn 3+5, Haxeltine & Prentice 1996.

Terrarium.compute_LAI — Method

compute_LAI(phenol, LAI_b)

Computes LAI, based on the balanced Leaf Area Index LAI_b:

Terrarium.compute_NPP — Method

compute_NPP(autoresp, GPP, Ra)

Computes Net Primary Productivity NPP as the difference between Gross Primary Production GPP and autotrophic respiration Ra in [kgC/m²/day].

Terrarium.compute_PAR — Method

compute_PAR(photo, swdown)

Computes NET Photosynthetically Active Radiation PAR [mol/m²/day].

Terrarium.compute_Ra — Method

compute_Ra(
    autoresp,
    vegcarbon_dynamics,
    T_air,
    Rd,
    phen,
    C_veg,
    GPP
)

Computes autotrophic respiration Ra as the sum of maintenance respiration Rm and growth respiration Rg in [kgC/m²/day].

Terrarium.compute_Rd — Method

compute_Rd(photo, Vc_max, β)

Computes the daily leaf respiration Rd [gC/m²/day], Eqn 10, Haxeltine & Prentice 1996 and Eq. 10 PALADYN (Willeit 2016).

Terrarium.compute_Rg — Method

compute_Rg(autoresp, GPP, Rm)

Computes growth respiration Rg in [kgC/m²/day].

Terrarium.compute_Rm — Method

compute_Rm(
    autoresp,
    vegcarbon_dynamics,
    T_air,
    Rd,
    phen,
    C_veg
)

Computes maintenance respiration Rm in [kgC/m²/day].

Terrarium.compute_Vc_max — Method

compute_Vc_max(
    photo,
    c_1,
    APAR,
    Kc,
    Ko,
    Γ_star,
    pres_i,
    pres_O2
)

Computes the maximum daily rate of net photosynthesis Vc_max [gC/m²/day], following the coordination hypothesis (acclimation), see Harrison 2021 Box 2. Note: this is not the same formula in PALADYN paper, this implementaion is taken from the code

Terrarium.compute_auxiliary! — Function

compute_auxiliary!(state, model::AbstractModel)

Computes updates to all auxiliary variables based on the current prognostic state of the model.

Terrarium.compute_auxiliary! — Method

Updates any state variables associated with the given boundary conditions.

Terrarium.compute_c1_c2 — Method

compute_c1_c2(photo, T_air, Γ_star, Kc, Ko, pres_i, pres_O2)

Computes factor for light-limited assimilation c_1 and factor for RuBisCO-limited assimilation c_2, Eqs. C4+C5, PALADYN (Willeit 2016).

Terrarium.compute_f_deciduous — Method

compute_f_deciduous(phenol)

Computes f_deciduous, a factor for smooth transition between evergreen and deciduous [-].

Terrarium.compute_phen — Method

compute_phen(phenol)

Computes phen, the phenology factor [-].

Terrarium.compute_photosynthesis — Method

compute_photosynthesis(
    photo,
    T_air,
    swdown,
    pres,
    co2,
    LAI,
    λc
)

Computes Gross Primary Production GPPin [kgC/m²/day] and leaf respiration Rd in [gC/m²/day]

Terrarium.compute_pres_i — Method

compute_pres_i(photo, λc, pres_a)

Computes intercellular CO2 partial pressure [Pa], Eq. 67, PALADYN (Willeit 2016).

Terrarium.compute_resp10 — Method

compute_resp10(autoresp)

Computes resp10

Terrarium.compute_tendencies! — Function

compute_tendencies!(state, model::AbstractModel)

Computes tendencies for all prognostic state variables for model stored in the given state.

Terrarium.compute_tendencies! — Method

Computes any tendency contributions from the given boundary conditions.

Terrarium.compute_vpd — Method

compute_vpd(T_air, q_air, pres)

Computes the vapor pressure deficit from air temperature, specific humidity, and surface pressure.

Terrarium.compute_Γ_star — Method

compute_Γ_star(photo, τ, pres_O2)

Computes the CO2 compensation point Γ_star, Eq. C6, PALADYN (Willeit 2016).

Terrarium.compute_Λ_loc — Method

compute_Λ_loc(vegcarbon_dynamics, LAI_b)

Computes the local litterfall rate Λ_loc based on the balanced Leaf Area Index LAI_b (assuming evergreen PFTs), Eq. 75, PALADYN (Willeit 2016).

Terrarium.compute_β — Method

compute_β(photo)

Computes the soil-moisture limiting factor β, Eq. 66, PALADYN (Willeit 2016).

Terrarium.compute_γv — Method

compute_γv(veg_dynamics)

Computes the disturbance rateγv, Eq. 80, PALADYN (Willeit 2016).

Terrarium.compute_λ_NPP — Method

compute_λ_NPP(vegcarbon_dynamics, LAI_b)

Computes λ_NPP,a factor determining the partitioning of NPP between increase of vegetation carbon of the existing vegetated area and spreading of the given PFT based on the balanced Leaf Area Index LAI_b, Eq. 74, PALADYN (Willeit 2016).

Terrarium.compute_λc — Method

compute_λc(stomcond, vpd)

Computes the ratio of leaf-internal and air CO2 concentration λc, derived from the optimal stomatal conductance model (Medlyn et al. 2011), Eq. 71, PALADYN (Willeit 2016).

Terrarium.compute_ν_star — Method

compute_ν_star(veg_dynamics, ν)

Computes ν_star which is the maximum between the current vegetation fraction ν and the seed fraction ν_seed [-], to ensure that a PFT is always seeded.

Terrarium.compute_ν_tend — Method

compute_ν_tend(
    veg_dynamics,
    vegcarbon_dynamics,
    LAI_b,
    C_veg,
    ν
)

Computes the vegetation fraction tendency for a single PFT, Eq. 73, PALADYN (Willeit 2016).

Terrarium.convert_dt — Method

convert_dt(dt)

Convert dts of type Period to a numeric value in seconds. Return dt if already a number.

Terrarium.fastiterate — Method

fastiterate(f!::F, iters::NTuple{N,Any}...) where {F,N}

Same as fastmap but simply invokes f! on each argument set without constructing a tuple.

Terrarium.fastmap — Method

fastmap(f::F, iter::NamedTuple...) where {F}

Same as map for NamedTuples but with guaranteed type stability. fastmap is a @generated function which unrolls calls to f into a loop-free tuple construction expression. All named tuples must have the same keys but in no particular order. The returned NamedTuple

Terrarium.fastmap — Method

fastmap(f::F, iter::NTuple{N,Any}...) where {F,N}

Same as map for NTuples but with guaranteed type stability. fastmap is a @generated function which unrolls calls to f into a loop-free tuple construction expression.

Terrarium.get_boundary_conditions — Method

get_boundary_conditions(model::AbstractModel)::AbstractBoundaryConditions

Returns the boundary conditions associated with this model.

Terrarium.get_dt — Function

get_dt(timestepper::AbstractTimeStepper)

Get the current timestep size for the time stepper.

Terrarium.get_field_boundary_conditions — Method

get_field_boundary_conditions(_, grid)

Constructs a suitable Oceananigans BoundaryCondition for the given state variable var on grid. If bcs is a NamedTuple, it is assumed that the keys correspond to variable names and this method is invoked recursively on the entry matching the name of var, if it exists. Otherwise, nothing is returned.

Terrarium.get_field_grid — Function

get_field_grid(grid::AbstractLandGrid)::Oceananigans.AbstractGrid

Returns the underlying Oceananigans grid type for Fields defined on the given land grid.

Terrarium.get_field_initializers — Method

Returns the field initializers stored in init.vars merged with the field initializers defined by the inner initializer. The field initilaizers defined by the given FieldInitializers take precedence in the merge.

Terrarium.get_field_initializers — Method

get_field_initializers(_)

Returns a named tuple of Oceananigans Field initializer functions where the keys correspond to the names of the respective state variables.

Terrarium.get_grid — Method

get_grid(model::AbstractModel)::AbstractLandGrid

Return the spatial grid associated with this model.

Terrarium.get_initializer — Method

get_initializer(model::AbstractModel)::AbstractInitializer

Returns the initializer associated with this model.

Terrarium.get_npoints — Method

get_npoints(spacing)

Return the number of vertical layers defined by this discretization.

Terrarium.get_spacing — Method

get_spacing(spacing)

Return a Vector of vertical layer thicknesses according to the given discretization.

Terrarium.get_time_stepping — Method

get_time_stepping(model::AbstractModel)::AbstractTimeStepper

Returns the time stepping scheme associated with this model.

Terrarium.getvar — Function

getvar(::AbstractClosureRelation, dims::VarDims)

Returns an AuxiliaryVariable corresponding to the closure variable defined by the given closure relation.

Terrarium.heatcapacity — Method

heatcapacity(props, fracs)

Computes the bulk heat capacity of a finite volume from the given volumetric fractions.

Terrarium.inferdims — Method

inferdims(_)

Complement of inferloc that infers the appropriate VarDims from the given Field.

Terrarium.inferloc — Method

inferloc(_)

Retrieves the Oceananigans field "location" for the given variable dimensions. The location refers to the where the variable is defined on a staggered finite volume grid: Center refers to the grid cell centroid, Face to the boundary, and Nothing to a quantity that has no dimensionality in that direction. Currently, we assume all variables to be defined on grid cell centers, but this could be relaxed later if necessary by modifying or extending VarDims.

Terrarium.initialize! — Function

initialize!(state, model::AbstractModel)
initialize!(state, model::AbstractModel, initializer::AbstractInitializer)

Calls initialize! on the model and its corresponding initializer. This method only needs to be implemented if initialization routines are necessary in addition to direct field/variable initializers.

Terrarium.initialize! — Method

Resets the simulation clock and calls initialize!(state, model) on the underlying model which should reset all state variables to their values as defiend by the model initializer.

Terrarium.initialize! — Method

Terrarium.initialize — Method

Creates and initializes a ModelState for the given model with the given clock state. This method allocates all necessary Fields for the state variables and calls initialize!(::ModelState). Note that this method is not type stable and should not be called in an Enzyme autodiff call.

Terrarium.input — Method

input(name, dims; units, desc)

Convenience constructor method for InputVariable.

Terrarium.is_adaptive — Function

is_adaptive(timestepper::AbstractTimeStepper)

Returns true if the given time stepper is adaptive, false otherwise.

Terrarium.merge_duplicates — Method

merge_duplicates(values)

Filter out duplicates from the given tuple. Note that this method is not type stable or allocation-free!

Terrarium.merge_recursive — Method

merge_recursive(nt1::NamedTuple, nt2::NamedTuple)

Recursively merge two nested named tuples. This implementation is loosely based on the one in NamedTupleTools authored by Jeffrey Sarnoff.

Terrarium.namespace — Method

namespace(name, vars)

Convenience constructor method for variable Namespaces.

Terrarium.organic_fraction — Method

organic_fraction(idx, state, bgc::ConstantSoilCarbonDenisty)

Calculate the soil organic carbon fraction at the given grid index. For ConstantSoilCarbonDenisty, this simply returns a constant parameter. Implementations of soil carbon dynamics would want to compute this based on the prognostic state of the carbon content/pool stored in the soil.

Terrarium.organic_fraction — Method

organic_fraction(bgc::ConstantSoilCarbonDenisty)

Calculate the organic solid fraction based on the prescribed SOC and natural porosity/density of the organic material.

Terrarium.organic_porosity — Method

organic_porosity(bgc::ConstantSoilCarbonDenisty)

Get the prescribed natural porosity of organic soil.

Terrarium.partial_pressure_CO2 — Method

partial_pressure_CO2(pres, conc_co2)

Compute partial pressure of CO2 from surface pressure and CO2 concentration in Pa.

Terrarium.partial_pressure_O2 — Method

partial_pressure_O2(pres)

Compute partial pressure of oxygen from surface pressure in Pa.

Terrarium.piecewise_linear — Method

piecewise_linear(knots; extrapolation)

Return a function f(z) that linearly interpolates between the given knots.

Terrarium.prognostic — Method

prognostic(name, dims; units, desc)

Convenience constructor method for PrognosticVariable.

Terrarium.safediv — Method

safediv(x, y)

Evaluate x / y unless iszero(y) is true, then returns zero.

Terrarium.tendency — Method

tendency(progname, progdims, progunits)

Creates an AuxiliaryVariable for the tendency of a prognostic variable with the given name, dimensions, and physical units. This constructor is primarily used internally by other constructors and does not usually need to be called by implementations of variables.

Terrarium.thermalconductivity — Method

thermalconductivity(props, fracs)

Computes the bulk thermal conductivity of a finite volume from the given volumetric fractions.

Terrarium.timestep! — Function

timestep!(state, model::AbstractModel, timestepper::AbstractTimeStepper, [dt = nothing])

Advance the model state by one time step, or by dt units of time.

Terrarium.timestep! — Method

Convenience dispatch for timestep! that forwards to timestep!(state, model, get_time_stepping(model), dt).

Terrarium.timestep! — Method

timestep!(state)

Advance the model forward by one timestep with optional timestep size dt.

Terrarium.tuplejoin — Method

tuplejoin()

Concatenate one or more tuples together.

Terrarium.update_inputs! — Method

update_inputs!(_, _, _)

Updates the values of input variables stored in fields from the given input source. Default implementation simply returns nothing.

Terrarium.vardims — Method

vardims(var)

Retrieves the grid dimensions on which this variable is defined.

Terrarium.variables — Function

variables(model::AbstractModel)

Return a Tuple of AbstractVariables (i.e. PrognosticVariable, AuxiliaryVariable, etc.) defined by the model. For models that consist of one or more sub-models, variables may optionally be grouped into namespaces by returning a NamedTuple where the keys correspond to the group names.

Terrarium.variables — Method

Like models/processes, boundary conditions can define state variables which may be computed from other state variables or from input data in compute_auxiliary!.

Terrarium.variables — Method

Alias for Variables(vars...)

Terrarium.varname — Method

varname(var)

Retrieves the name of the given variable or closure. For closure relations, varname should return the name of the variable returned by the closure relation.

Terrarium.varunits — Method

varunits(var)

Retrieves the physical units for the given variable.